requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.109.0
python-dotenv==1.0.0
click==8.1.7
reportlab==4.0.9
beautifulsoup4==4.12.2
html2text==2020.1.16
urllib3==2.1.0
flask==3.0.0